Add some basic e2e tests
I've gone for Cypress, as it runs in a real
browser and seems the most friction-free way
to run e2e tests. Nest comes bundled with
`supertest`, which seems more akin to testing
APIs, rather than web applications. I've also
tweaked the DB config, as it was set up
assuming we were going to use the DB in the
unit tests, which I think we should try not to do
to keep them speedy