Use a more "Nest-like" way to validate
Instead of having to validate each time, we're using
a custom Validation error (which includes the full
validation errors in a format we can work in) and
catching that error in an exception filter. We can
then register the filters and tell them what template
to render and what object to send to the view if
there is an error.